+++ lib/po_editor_downloader.dart
@@ -1,6 +1,5 @@
 library;
 
-export 'src/argument_value_parser.dart';
 export 'src/config_reader.dart';
 export 'src/language.dart';
 export 'src/po_editor_config.dart';